worker2/cycler/requirements.txt
pyaes
mysql-connector-python
confluent-kafka
kafka-python
prometheus-client
